refactor: move RollingOnDemandStatsProducer into moskito-core
RollingOnDemandStatsProducer was the sole class in the
moskito-additional-producers module. It's a general-purpose producer
(an LRU-bounded alternative to OnDemandStatsProducer) that belongs
alongside its siblings in moskito-core's net.anotheria.moskito.core.dynamic
package (next to OnDemandStatsProducer and
EntryCountLimitedOnDemandStatsProducer).

- Move the class (and its test) core.dynamic; drop the now same-package
  imports of OnDemandStatsProducer / IOnDemandStatsFactory.
- Add its only non-core dependency (concurrentlinkedhashmap-lru) to
  moskito-core.
- Remove the now-empty moskito-additional-producers module and its
  reactor entry.

Breaking change (fits the 5.0 major): the class moves from
net.anotheria.moskito.extensions.producers to
net.anotheria.moskito.core.dynamic, and the moskito-additional-producers
artifact is gone — consumers get the class from moskito-core instead.

Verified: moved test passes in core; full reactor builds.
