--- moskito-extensions/moskito-additional-producers/pom.xml
@@ -1,37 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <parent>
-	<groupId>net.anotheria</groupId>
-	<artifactId>moskito</artifactId>
-	<version>5.0.0-SNAPSHOT</version>
-	<relativePath>../../pom.xml</relativePath>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>moskito-additional-producers</artifactId>
-  <version>5.0.0-SNAPSHOT</version>
-  <name>additional producers</name>
-
-  <dependencies>
-      <dependency>
-		<groupId>net.anotheria</groupId>
-		<artifactId>moskito-core</artifactId>
-		<version>${project.version}</version>
-	</dependency>
-      <dependency>
-          <groupId>org.slf4j</groupId>
-          <artifactId>slf4j-api</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>ch.qos.logback</groupId>
-          <artifactId>logback-classic</artifactId>
-          <!--scope>test</scope-->
-      </dependency>
-
-      <dependency>
-          <groupId>com.googlecode.concurrentlinkedhashmap</groupId>
-          <artifactId>concurrentlinkedhashmap-lru</artifactId>
-          <version>1.4.2</version>
-      </dependency>
-  </dependencies>
-</project>
