+++ moskito-extensions/pom.xml
@@ -14,7 +14,6 @@
 
 	<modules>
 		<module>moskito-notification-providers</module>
-		<module>moskito-additional-producers</module>
 		<module>moskito-sampling-api</module>
         <module>moskito-disk-space-monitoring</module>
         <module>moskito-mongodb</module>
